+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 * Add platform verification id field to the VerifiedName model
+* Integrate platform verification id into app
 
 [2.4.0] - 2024-04-23
 ~~~~~~~~~~~~~~~~~~~~